sandbox/doc-agent/Doc_Service_Agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-docx==1.1.0
pydantic==2.6.4
